docs/sample-project-2-gas-fee-monitor/requirements.txt
# Core dependencies
uagents>=0.6.0
requests>=2.28.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
matplotlib>=3.7.0  # For optional visualization features
